refactor: move to a separate pass for maintainence
