+++ src/enzyme_ad/jax/Utils.h
@@ -470,6 +470,12 @@ static arith::CmpIPredicate swapPredicate(arith::CmpIPredicate pred) {
   llvm_unreachable("unknown cmpi predicate kind");
 }
 
+SmallVector<int64_t>
+findReshapeInsertionDims(RankedTensorType inputType,
+                         RankedTensorType outputType);
+
+bool isOnlyUsedInOperation(Operation *operation, Operation *parentOp);
+
 } // namespace enzyme
 
 namespace stablehlo {
